[MacOS] ListView - further sharing
