+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/ListViewBaseInternalContainer.macOS.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using AppKit;
+
+namespace Windows.UI.Xaml.Controls
+{
+	internal partial class ListViewBaseInternalContainer : NSCollectionViewItem
+	{
+	}
+}
