+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/ListViewBaseScrollContentPresenter.macOS.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Windows.Foundation;
+
+namespace Windows.UI.Xaml.Controls
+{
+	public sealed partial class ListViewBaseScrollContentPresenter
+	{
+		public Rect MakeVisible(UIElement visual, Rect rectangle) => rectangle; //TODO: ? (copied from iOS)
+	}
+}
