+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/ListViewBaseSource.iOS.cs
@@ -69,68 +69,7 @@ namespace Windows.UI.Xaml.Controls
 		{
 		}
 
-		#region Properties
-
-		#endregion
-
-		public ListViewBaseSource(NativeListViewBase owner)
-		{
-			Owner = owner;
-		}
-
 		#region Overrides
-		public override nint NumberOfSections(UICollectionView collectionView)
-		{
-			var itemsSections = Owner.XamlParent.IsGrouping ?
-				Owner.XamlParent.NumberOfDisplayGroups :
-				1;
-			return itemsSections + SupplementarySections;
-		}
-
-		public override nint GetItemsCount(UICollectionView collectionView, nint section)
-		{
-			int count;
-			if (Owner.XamlParent.IsGrouping)
-			{
-				count = GetGroupedItemsCount(section);
-			}
-			else
-			{
-				count = GetUngroupedItemsCount(section);
-			}
-
-			if (this.Log().IsEnabled(Microsoft.Extensions.Logging.LogLevel.Debug))
-			{
-				this.Log().Debug($"Count requested for section {section}, returning {count}");
-			}
-			return count;
-		}
-
-		private int GetUngroupedItemsCount(nint section)
-		{
-			int count;
-			if (section == 0)
-			{
-				count = Owner.XamlParent.NumberOfItems;
-			}
-			else
-			{
-				// Extra section added to accommodate header+footer, contains no items.
-				count = 0;
-			}
-
-			return count;
-		}
-
-		private int GetGroupedItemsCount(nint section)
-		{
-			if ((int)section >= Owner.XamlParent.NumberOfDisplayGroups)
-			{
-				// Header+footer section which is empty
-				return 0;
-			}
-			return Owner.XamlParent.GetDisplayGroupCount((int)section);
-		}
 
 		public override void CellDisplayingEnded(UICollectionView collectionView, UICollectionViewCell cell, NSIndexPath indexPath)
 		{
@@ -260,98 +199,6 @@ namespace Windows.UI.Xaml.Controls
 			{
 				this.Log().LogDebug($"WillDisplayCell for cell at {indexPath}");
 			}
-		}
-
-		public override UICollectionReusableView GetViewForSupplementaryElement(
-			UICollectionView collectionView,
-			NSString elementKind,
-			NSIndexPath indexPath)
-		{
-			var listView = (NativeListViewBase)collectionView;
-
-			if (elementKind == NativeListViewBase.ListViewHeaderElementKind)
-			{
-				return GetBindableSupplementaryView(
-					collectionView: collectionView,
-					elementKind: NativeListViewBase.ListViewHeaderElementKindNS,
-					indexPath: indexPath,
-					reuseIdentifier: NativeListViewBase.ListViewHeaderReuseIdentifierNS,
-					context: listView.Header,
-					template: listView.HeaderTemplate,
-					style: null
-				);
-			}
-
-			else if (elementKind == NativeListViewBase.ListViewFooterElementKind)
-			{
-				return GetBindableSupplementaryView(
-					collectionView: collectionView,
-					elementKind: NativeListViewBase.ListViewFooterElementKindNS,
-					indexPath: indexPath,
-					reuseIdentifier: NativeListViewBase.ListViewFooterReuseIdentifierNS,
-					context: listView.Footer,
-					template: listView.FooterTemplate,
-					style: null
-				);
-			}
-
-			else if (elementKind == NativeListViewBase.ListViewSectionHeaderElementKind)
-			{
-				// Ensure correct template can be retrieved
-				UpdateLastMaterializedItem(indexPath);
-
-				return GetBindableSupplementaryView(
-					collectionView: collectionView,
-					elementKind: NativeListViewBase.ListViewSectionHeaderElementKindNS,
-					indexPath: indexPath,
-					reuseIdentifier: NativeListViewBase.ListViewSectionHeaderReuseIdentifierNS,
-					//ICollectionViewGroup.Group is used as context for sectionHeader
-					context: listView.XamlParent.GetGroupAtDisplaySection(indexPath.Section).Group,
-					template: GetTemplateForGroupHeader(indexPath.Section),
-					style: listView.GroupStyle?.HeaderContainerStyle
-				);
-			}
-
-			else
-			{
-				throw new NotSupportedException("Unsupported element kind: {0}".InvariantCultureFormat(elementKind));
-			}
-		}
-
-		private UICollectionReusableView GetBindableSupplementaryView(
-			UICollectionView collectionView,
-			NSString elementKind,
-			NSIndexPath indexPath,
-			NSString reuseIdentifier,
-			object context,
-			DataTemplate template,
-			Style style)
-		{
-			var supplementaryView = (ListViewBaseInternalContainer)collectionView.DequeueReusableSupplementaryView(
-				elementKind,
-				reuseIdentifier,
-				indexPath);
-
-			using (supplementaryView.InterceptSetNeedsLayout())
-			{
-				if (supplementaryView.Content == null)
-				{
-					supplementaryView.Owner = Owner;
-					var content = CreateContainerForElementKind(elementKind);
-					content.HorizontalContentAlignment = HorizontalAlignment.Stretch;
-					content.VerticalContentAlignment = VerticalAlignment.Stretch;
-					supplementaryView.Content = content
-						.Binding("Content", "");
-				}
-				supplementaryView.Content.ContentTemplate = template;
-				supplementaryView.Content.DataContext = context;
-				if (style != null)
-				{
-					supplementaryView.Content.Style = style;
-				}
-			}
-
-			return supplementaryView;
 		}
 
 		internal void SetIsAnimatedScrolling() => _isInAnimatedScroll = true;
