+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/NativeListViewBase.macOS.cs
@@ -3,23 +3,26 @@ using System.Collections.Generic;
 using System.Text;
 using AppKit;
 using CoreGraphics;
+using Uno;
 using Windows.UI.Xaml.Automation.Peers;
 
 namespace Windows.UI.Xaml.Controls
 {
 	public partial class NativeListViewBase : NSCollectionView
 	{
-		internal CGPoint ContentOffset
-		{
-			get => throw new NotImplementedException();
-			set => throw new NotImplementedException();
-		}
+		[NotImplemented]
+		internal CGPoint ContentOffset { get; set; } // TODO: implement
 
 		internal ListViewBaseSource Source
 		{
-			get => throw new NotImplementedException();
-			set => throw new NotImplementedException();
+			get => base.DataSource as ListViewBaseSource;
+			set => base.DataSource = value;
 		}
+		[NotImplemented]
+		public ScrollBarVisibility HorizontalScrollBarVisibility { get; internal set; } //TODO: implement (https://stackoverflow.com/questions/49724516/how-to-hide-nscollectionview-scroll-indicator/54062881 ?)
+		[NotImplemented]
+		public ScrollBarVisibility VerticalScrollBarVisibility { get; internal set; } //TODO: implement
+
 		public AutomationPeer GetAutomationPeer()
 		{
 			return null;
